refactor: move jubilant tests into jubilant/ subdirectory
- Create tests/integration/jubilant/ with its own conftest.py that owns
  traefik_charm, juju, and no-op shadows for the two autouse fixtures
  (copy_traefik_library_into_tester_charms, setup_env)
- Move test_dynamic_configs_jubilant.py and test_tls_multi_unit_jubilant.py
  into the new subdirectory; remove their now-redundant local traefik_charm
  fixtures
- Clean up tests/integration/conftest.py: remove ops_test override, remove
  jubilant name-guards, remove juju_fixture, remove ops_test is None check
- Add a second integration-suite in spread.yaml for tests/integration/jubilant/
  so charm-ci auto-discovers jubilant tests independently; move juju4 variant
  overrides into that suite
