+++ tests/integration/jubilant/test_tls_multi_unit_jubilant.py
@@ -6,7 +6,6 @@
 
 import json
 import logging
-import os
 from pathlib import Path
 
 import jubilant
@@ -32,20 +31,6 @@ NUM_TRAEFIK_UNITS = 2
 # ---------------------------------------------------------------------------
 # Fixtures
 # ---------------------------------------------------------------------------
-@pytest.fixture(scope="module")
-def traefik_charm():
-    charm_path = os.environ.get("CHARM_PATH")
-    if charm_path:
-        return Path(charm_path).resolve()
-    charms = sorted(Path(".").glob("traefik*.charm"))
-    if charms:
-        return charms[0]
-    raise FileNotFoundError(
-        "Set CHARM_PATH to the built traefik charm, "
-        "or place a traefik*.charm file in the repo root."
-    )
-
-
 @pytest.fixture(scope="module", name="traefik_app")
 def traefik_fixture(juju, traefik_charm):
     """Deploy traefik."""
